requirements.txt
fastapi==0.115.11
google_cloud_aiplatform==1.71.1
gradio==5.20.1
numpy==2.2.3
pandas==2.2.3
uvicorn==0.34.0
vertexai==1.71.1
jinja2
re